test SEQUENCE EPHEMERAL

+++ shoebox/test/com/keepit/common/zookeeper/ZooKeeperClientTest.scala
@@ -12,14 +12,14 @@ import org.specs2.mutable.Specification
 import com.keepit.common.db._
 import com.keepit.common.db.slick._
 import org.apache.zookeeper.CreateMode
+import org.apache.zookeeper.CreateMode._
 import scala.util.{Random, Try}
 
 class ZooKeeperClientTest extends Specification {
 
-  args(skipAll = true)
+  // args(skipAll = true)
 
-  def withClient[T](block: ZooKeeperClient => T): T = {
-    val path = Path("/test" + Random.nextLong.abs)
+  def withClient[T](block: ZooKeeperClient => T)(implicit path: Path = Path("/test" + Random.nextLong.abs), cleanup: Boolean = true): T = {
     println(s"starting test with root path $path")
     val zk = new ZooKeeperClient("localhost", 2000, path,
                     Some({zk1 => println(s"in callback, got $zk1")}))
@@ -27,7 +27,7 @@ class ZooKeeperClientTest extends Specification {
       zk.createPath(path)
       block(zk)
     } finally {
-      Try { zk.deleteRecursive(path) }
+      if (cleanup) Try { zk.deleteRecursive(path) }
       zk.close
     }
   }
@@ -48,7 +48,7 @@ class ZooKeeperClientTest extends Specification {
 
     "connect to server and set some data" in {
       withClient { zk =>
-        zk.createNode(Node("/testNode"), "foo".getBytes, CreateMode.PERSISTENT)
+        zk.createNode(Node("/testNode"), "foo".getBytes, PERSISTENT)
         zk.watchNode(Node("/testNode"), { (data : Option[Array[Byte]]) =>
           data match {
             case Some(d) => println("Data updated: %s".format(new String(d)))
@@ -63,32 +63,55 @@ class ZooKeeperClientTest extends Specification {
 
     "monitor node children" in {
       withClient { zk =>
-        zk.create(Path("/parent"), null, CreateMode.PERSISTENT)
+        zk.create(Path("/parent"), null, PERSISTENT)
         zk.watchChildren(Path("/parent"), { (children : Seq[Node]) =>
           println("Children: %s".format(children.mkString(", ")))
         })
-        zk.createNode(Node("/parent/child1"), null, CreateMode.PERSISTENT)
-        zk.createNode(Node("/parent/child2"), null, CreateMode.PERSISTENT)
+        zk.createNode(Node("/parent/child1"), null, PERSISTENT)
+        zk.createNode(Node("/parent/child2"), null, PERSISTENT)
         zk.deleteNode(Node("/parent/child1"))
-        zk.createNode(Node("/parent/child3"), null, CreateMode.PERSISTENT)
+        zk.createNode(Node("/parent/child3"), null, PERSISTENT)
         zk.deleteRecursive(Path("/parent"))
       }
+    }
+
+    "SEQUENCE EPHEMERAL (Service Instances) nodes" in {
+      var basePath = Path("/test" + Random.nextLong.abs)
+      withClient { zk =>
+        basePath = zk.basePath
+        val parent = Path("/parent")
+        zk.create(parent, null, PERSISTENT)
+        zk.watchChildren(parent, { (children : Seq[Node]) =>
+          println("Service Instances: %s".format(children.mkString(", ")))
+        })
+        println(zk.createNode(Node("/parent/child1"), null, EPHEMERAL_SEQUENTIAL))
+        println(zk.createNode(Node("/parent/child2"), null, EPHEMERAL_SEQUENTIAL))
+        println(zk.createNode(Node("/parent/child3"), null, EPHEMERAL_SEQUENTIAL))
+        zk.getChildren(parent).size === 3
+
+        zk.create(Path("other"), null, PERSISTENT)
+        println(zk.createNode(Node("/other/child"), null, PERSISTENT))
+      }(basePath, false)
+      withClient { zk =>
+        zk.getChildren(Path("/other")).size === 1
+        zk.getChildren(Path("/parent")).size === 0
+      }(basePath, true)
     }
 
     "For a given node, automatically maintain a map from the node's children to the each child's data" in {
       withClient { zk =>
         val childMap = collection.mutable.Map[Node, String]()
 
-        zk.create(Path("/parent"), null, CreateMode.PERSISTENT)
+        zk.create(Path("/parent"), null, PERSISTENT)
         zk.watchChildrenWithData(Path("/parent"), childMap, {data => new String(data)})
 
-        zk.create(Path("/parent/a"), "foo".getBytes, CreateMode.PERSISTENT)
-        zk.create(Path("/parent/b"), "bar".getBytes, CreateMode.PERSISTENT)
+        zk.create(Path("/parent/a"), "foo".getBytes, PERSISTENT)
+        zk.create(Path("/parent/b"), "bar".getBytes, PERSISTENT)
         println("child map: %s".format(childMap)) // NOTE: real code should synchronize access on childMap
 
         zk.delete(Path("/parent/a"))
         zk.set(Node("/parent/b"), "bar2".getBytes)
-        zk.createNode(Node("/parent/c"), "baz".getBytes, CreateMode.PERSISTENT)
+        zk.createNode(Node("/parent/c"), "baz".getBytes, PERSISTENT)
         println("child map: %s".format(childMap)) // NOTE: real code should synchronize access on childMap
       }
     }
